requirements.txt
autoenv==1.0.0
dj-database-url==0.3.0
dj-static==0.0.6
-e git+https://github.com/django/django.git@b8f9fd75910acdcbd7f485c4acc55f98a612d86c#egg=Django-dev
django-bootstrap3==5.1.1
django-toolbelt==0.0.1
gunicorn==19.2.1
psycopg2==2.6
static3==0.5.1
